+++ src/components/testIds.ts
@@ -1,6 +1,7 @@
 export const testIds = {
   appConfig: {
     container: 'data-testid ac-container',
+    useAzureOpenAI: 'data-testid ac-use-azure-openai',
     openAIKey: 'data-testid ac-openai-api-key',
     openAIOrganizationID: 'data-testid ac-openai-api-organization-id',
     openAIUrl: 'data-testid ac-openai-api-url',